requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.0
Flask-Cors==4.0.0
gcloud==0.17.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.3
oauth2client==3.0.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.4.3
pyparsing==3.1.1
Pyrebase==3.0.27
python-dotenv==1.0.0
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
werkzeug==3.0.1
zipp==3.17.0
